# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Poket Project API",
        "description": "API for managing user sets, see all cards and all sets of pokemon.",
        "contact": {
            "email": "hello@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cards/{card}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Get a card",
                "operationId": "dc004b7881adb0c6fac597bbe4b1e6fc",
                "parameters": [
                    {
                        "name": "card",
                        "in": "path",
                        "description": "Card identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Obtained Card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get the profile of the authenticated user",
                "operationId": "f1784685f8270527da017b3616288443",
                "responses": {
                    "200": {
                        "description": "User profile successfully obtained",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the profile of the authenticated user",
                "operationId": "875e8801180287e4193f1aced9023960",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete user account",
                "operationId": "696324677bc8a587bf01679e38cf9989",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "422": {
                        "description": "The password is incorrect"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "ef623afc428521756f486439e3307dc6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Incorrect credentials"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "8930e64d1f680ad30f44dfa5e1368387",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect validation data"
                    }
                }
            }
        },
...
        "/api/sets/{set}": {
            "get": {
                "tags": [
                    "Sets"
                ],
                "summary": "Get a specific set",
                "operationId": "6f2b978e63a1fcb041f19e83e5ed1160",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set identificator",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set obtained correctly",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Set"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Set not found"
                    }
                }
            }
        },
        "/api/sets/{set}/cards": {
            "get": {
                "tags": [
                    "Sets"
                ],
                "summary": "Get the cards from a set",
                "operationId": "5e0677de97783fab8cf7e8e3163d0ec5",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set identificator",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cards from the set obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Set not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Card": {
                "properties": {
                    "id": {
                        "description": "Card identifier",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "supertype_id": {
                        "type": "integer"
                    },
                    "level": {
                        "type": "string",
                        "nullable": true
                    },
                    "hp": {
                        "type": "string",
                        "nullable": true
                    },
                    "evolvesFrom": {
                        "type": "string",
                        "nullable": true
                    },
                    "evolvesTo": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "rules": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "ancientTrait": {
                        "type": "string",
                        "nullable": true
                    },
                    "abilities": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "attacks": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "cost": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "name": {
                                    "type": "string"
                                },
                                "text": {
                                    "type": "string"
                                },
                                "damage": {
                                    "type": "string"
                                },
                                "convertedEnergyCost": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "weaknesses": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "value": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "resistances": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "retreatCost": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "convertedRetreatCost": {
                        "type": "integer"
                    },
                    "set_id": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    },
                    "artist": {
                        "type": "string"
                    },
                    "rarity_id": {
                        "type": "integer"
                    },
                    "flavorText": {
                        "type": "string"
                    },
                    "nationalPokedexNumbers": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "legalities": {
                        "properties": {
                            "unlimited": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "regulationMark": {
                        "type": "string",
                        "nullable": true
                    },
                    "images": {
                        "properties": {
                            "large": {
                                "type": "string"
                            },
                            "small": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "tcgplayer": {
                        "properties": {
                            "url": {
                                "type": "string"
                            },
                            "prices": {
                                "properties": {
                                    "holofoil": {
                                        "properties": {
                                            "low": {
                                                "type": "number"
                                            },
                                            "mid": {
                                                "type": "number"
                                            },
                                            "high": {
                                                "type": "number"
                                            },
                                            "market": {
                                                "type": "number"
                                            },
                                            "directLow": {
                                                "type": "number",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    },
                    "cardmarket": {
                        "properties": {
                            "url": {
                                "type": "string"
                            },
                            "prices": {
                                "properties": {
                                    "avg1": {
                                        "type": "number"
                                    },
                                    "avg7": {
                                        "type": "number"
                                    },
                                    "avg30": {
                                        "type": "number"
                                    },
                                    "lowPrice": {
                                        "type": "number"
                                    },
                                    "trendPrice": {
                                        "type": "number"
                                    },
                                    "germanProLow": {
                                        "type": "number"
                                    },
                                    "lowPriceExPlus": {
                                        "type": "number"
                                    },
                                    "reverseHoloLow": {
                                        "type": "number"
                                    },
                                    "suggestedPrice": {
                                        "type": "number"
                                    },
                                    "reverseHoloAvg1": {
                                        "type": "number"
                                    },
                                    "reverseHoloAvg7": {
                                        "type": "number"
                                    },
                                    "reverseHoloSell": {
                                        "type": "number"
                                    },
                                    "averageSellPrice": {
                                        "type": "number"
                                    },
                                    "reverseHoloAvg30": {
                                        "type": "number"
                                    },
                                    "reverseHoloTrend": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}